study-tools/listening-practice/requirements.txt
streamlit
youtube-transcript-api
pandas
python-dotenv
openai
pydub